The Great Migration of code from terminalView to tabsView
